# Lazy payload building