--- a/.travis.yml
+++ b/.travis.yml
@@ -3,3 +3,8 @@ language: go
 install: go get -t
 
 script: make test
+
+sudo: false
+
+go:
+    - tip