--- a/.travis.yml
+++ b/.travis.yml
@@ -7,4 +7,4 @@ script: make test
 sudo: false
 
 go:
-    - tip
+    - 1.5.2