--- a/command.go
+++ b/command.go
@@ -18,10 +18,10 @@ var (
 
 // Cmd is a redis command.
 type Cmd struct {
-	hash    uint64
-	payload []byte
-	res     interface{}
-	err     error
+	hash   uint64
+	fields []string
+	res    interface{}
+	err    error
 }
 
 // Build makes a command which will be send to the shard for 'key'. All
@@ -30,9 +30,9 @@ type Cmd struct {
 // simple command->reply ('WATCH').
 func Build(key string, fields ...string) *Cmd {
 	return &Cmd{
-		hash:    hashKey(key),
-		payload: buildCommand(fields, make([]byte, 0, 64)),
-		err:     ErrNotExecuted,
+		hash:   hashKey(key),
+		fields: fields,
+		err:    ErrNotExecuted,
 	}
 }
 
@@ -237,7 +237,7 @@ func resInt(x interface{}) (int, error) {
 	}
 }
 
-func buildCommand(fields []string, b []byte) []byte {
+func appendCommand(fields []string, b []byte) []byte {
 	b = append(b, '*')
 	b = strconv.AppendInt(b, int64(len(fields)), 10)
 	b = append(b, '\r', '\n')