--- a/command.go
+++ b/command.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"fmt"
 	"strconv"
-	"sync"
 )
 
 var (
@@ -34,27 +33,6 @@ func Build(key string, fields ...string) *Cmd {
 		fields: fields,
 		err:    ErrNotExecuted,
 	}
-}
-
-var poolCmds = sync.Pool{
-	New: func() interface{} {
-		return &Cmd{}
-	},
-}
-
-// PooledBuild is same as Build, but uses a sync.Pool. Use Cmd.Put() to put
-// them back in the queue.
-func PooledBuild(key string, fields ...string) *Cmd {
-	c := poolCmds.Get().(*Cmd)
-	c.hash = hashKey(key)
-	c.payload = buildCommand(fields, c.payload[:0])
-	c.err = ErrNotExecuted
-	return c
-}
-
-// Put puts a cmd back in the pool for reuse.
-func (c *Cmd) Put() {
-	poolCmds.Put(c)
 }
 
 // Get returns redis' result.