*Building benchmarks are better since we don't build the payload anymore, and Benchmark is slower since we now build the payload every run.

```
benchmark                old ns/op     new ns/op     delta
Benchmark                263837        303927        +15.19%
Benchmark-4              309077        296358        -4.12%
BenchmarkBuilding        423           205           -51.54%
BenchmarkBuilding-4      399           182           -54.39%
BenchmarkNoPrepare       29304         28059         -4.25%
BenchmarkNoPrepare-4     41190         29235         -29.02%

benchmark                old allocs     new allocs     delta
Benchmark                318            318            +0.00%
Benchmark-4              318            318            +0.00%
BenchmarkBuilding        2              2              +0.00%
BenchmarkBuilding-4      2              2              +0.00%
BenchmarkNoPrepare       11             11             +0.00%
BenchmarkNoPrepare-4     11             11             +0.00%

benchmark                old bytes     new bytes     delta
Benchmark                14161         14165         +0.03%
Benchmark-4              14161         14172         +0.08%
BenchmarkBuilding        128           128           +0.00%
BenchmarkBuilding-4      128           128           +0.00%
BenchmarkNoPrepare       448           432           -3.57%
BenchmarkNoPrepare-4     448           432           -3.57%
```