@alicebob 2016-01-08
I've merged all except the lazy payloads.